anomaly-worker/requirements.txt
confluent-kafka
scikit-learn
numpy
pymongo
psycopg2-binary
requests
python-dotenv
prometheus_client
psutil
psycopg2-binary
flask